refactoring working
