+++ src/krkn_lib/kubernetes/__init__.py
@@ -1,2 +1,1 @@
 from .krkn_kubernetes import *  # NOQA
-from .models import *  # NOQA
