+++ src/krkn_lib/models/telemetry/__init__.py
@@ -1,0 +1,1 @@
+from .models import *  # NOQA
