+++ src/krkn_lib/models/telemetry/models.py
