+++ src/krkn_lib/telemetry/__init__.py
@@ -1,2 +1,1 @@
 from .krkn_telemetry import *  # NOQA
-from .models import *  # NOQA
