+++ src/krkn_lib/tests/test_krkn_telemetry_models.py
@@ -1,7 +1,7 @@
 import json
 import unittest
 
-from krkn_lib.telemetry import ScenarioTelemetry, ChaosRunTelemetry
+from krkn_lib.models.telemetry import ScenarioTelemetry, ChaosRunTelemetry
 
 
 class KrknTelemetryModelsTests(unittest.TestCase):
